Moved contested solicitor create mid event handler to the same package as other solicitor create handlers
